+++ src/State.fs
@@ -1,7 +1,7 @@
 module State
 
 open Fetch
-open Thoth
+open Thoth.Json
 
 open Elmish
 open Fable.RemoteData
@@ -9,13 +9,14 @@ open Fable.RemoteData
 open Types
 
 module Cmds =
+    let baseUrl = "https://conduit.productionready.io/api/"
     let fetchArticles page =
         let articlesApi offset =
             promise {
-                let url = sprintf "https://conduit.productionready.io/api/articles?limit=10&offset=%i" offset
+                let url = sprintf "%sarticles?limit=10&offset=%i" baseUrl offset
                 let! response = fetch url []
                 let! txt = response.text()
-                let result = Json.Decode.Auto.fromString<ArticlesList>(txt, isCamelCase=true)
+                let result = Decode.Auto.fromString<ArticlesList>(txt, isCamelCase=true)
                 match result with
                 | Ok data ->
                     return Success data
@@ -25,12 +26,32 @@ module Cmds =
         let offset = page - 1
         Cmd.OfPromise.either articlesApi offset ArticlesFetched (Failure >> ArticlesFetched)
 
+    let fetchTags =
+
+        let tagsApi () =
+            promise {
+                let url = sprintf "%stags" baseUrl
+                let! response = fetch url []
+                let! txt = response.text()
+                let result = Decode.fromString Tag.ListDecoder txt
+                match result with
+                | Ok data ->
+                    return Success data
+                | Error err ->
+                    return Failure <| exn err
+            }
+        Cmd.OfPromise.either tagsApi () TagsFetched (Failure >> TagsFetched)
+
 let init() : Model * Cmd<Msg> =
     let initialPage = 1
     {
         Articles = Loading
+        PopularTags = Loading
         CurrentArticlesPage = initialPage
-    }, Cmds.fetchArticles initialPage
+    }, Cmd.batch [
+        Cmds.fetchArticles initialPage
+        Cmds.fetchTags
+    ]
 
 let update msg model : Model * Cmd<Msg> =
     match msg with
@@ -39,3 +60,6 @@ let update msg model : Model * Cmd<Msg> =
 
     | SetArticlesPage page ->
         { model with CurrentArticlesPage = page }, Cmds.fetchArticles page
+
+    | TagsFetched data ->
+        { model with PopularTags = data }, Cmd.none