+++ src/Types.fs
@@ -1,6 +1,7 @@
 module Types
 
 open Fable.RemoteData
+open Thoth.Json
 
 open System
 
@@ -11,12 +12,25 @@ type Author = {
     Following : bool
 }
 
+type Tag =
+    | Tag of string
+    static member Decoder : Decoder<Tag> =
+        Decode.object
+            (fun get ->
+                Tag <| get.Required.Raw Decode.string
+            )
+    static member ListDecoder : Decoder<Tag list> =
+        Decode.object
+            (fun get ->
+                get.Required.At ["tags"] (Decode.list Tag.Decoder)
+            )
+
 type Article = {
     Slug : string
     Title : string
     Description : string
     Body : string
-    TagList :string list
+    TagList : string list
     CreatedAt : DateTime
     UpdatedAt : DateTime
     Favorited : bool
@@ -31,9 +45,11 @@ type ArticlesList = {
 
 type Model = {
     Articles : RemoteData<exn, ArticlesList>
+    PopularTags : RemoteData<exn, Tag list>
     CurrentArticlesPage : int
 }
 
 type Msg =
     | ArticlesFetched of articles: RemoteData<exn, ArticlesList>
+    | TagsFetched of articles: RemoteData<exn, Tag list>
     | SetArticlesPage of int